+++ src/input.json
@@ -5,6 +5,11 @@
     {
       "elements": [
         {
+          "type" : "random-id",
+          "text" : "survey id",
+          "title": "Your survey id is"
+         },
+        {
           "type": "text",
           "isRequired": true,
           "name": "Age",
@@ -447,6 +452,8 @@
   ],
   "random": [
     { "for": [1, 7, 8, 9]},
+    { "for": [1,3] },
+    { "for": [1,3] },
     { "for": [1,3] }
   ]
 }
